feat: move converter.js functionality to index.js
